backend/requirements.txt
Django==3.2.12
djangorestframework==3.12.4
django-cors-headers==3.10.0
django-rest-auth==0.9.5
djangorestframework-simplejwt==5.0.0
